requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==4.10.1
Markdown==3.3.6
promise==2.3
pycodestyle==2.8.0
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tzdata==2021.5
zipp==3.7.0
